pyproject.toml
[project]
name = "anytype"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jinja2>=3.1.6",
    "python-dotenv>=1.1.1",
]
